✨ Add validate (#1000)
