+++ lamindb/__init__.py
@@ -147,7 +147,9 @@ if _INSTANCE_SETUP:
     from . import _label  # noqa
     from . import _registry  # noqa
     from . import _storage  # noqa
+    from . import _synonym  # noqa
     from . import _transform  # noqa
+    from . import _validate  # noqa
     from ._delete import delete  # noqa
     from ._registry import select_backward as select  # noqa
     from ._save import save  # noqa
