+++ lamindb/_registry.py
@@ -1,5 +1,5 @@
 import builtins
-from typing import Dict, Iterable, List, Literal, NamedTuple, Optional, Set, Union
+from typing import Dict, Iterable, List, NamedTuple, Optional, Union
 
 import pandas as pd
 from django.core.exceptions import FieldDoesNotExist
@@ -16,7 +16,7 @@ from lnschema_core.types import ListLike, StrField
 from lamindb.dev.utils import attach_func_to_class_method
 
 from . import _TESTING
-from ._from_values import _has_species_field, get_or_create_records
+from ._from_values import get_or_create_records
 from .dev._feature_manager import create_features_df
 from .dev._settings import settings
 
@@ -285,125 +285,6 @@ def lookup(cls, field: Optional[StrField] = None) -> NamedTuple:
     return _lookup(cls=cls, field=field)
 
 
-def _inspect(
-    cls,
-    identifiers: ListLike,
-    field: StrField,
-    *,
-    case_sensitive: bool = False,
-    inspect_synonyms: bool = True,
-    return_df: bool = False,
-    logging: bool = True,
-    **kwargs,
-) -> Union["pd.DataFrame", Dict[str, List[str]]]:
-    """{}"""
-    from lamin_utils._inspect import inspect
-
-    if not isinstance(field, str):
-        field = field.field.name
-
-    cls = cls.model if isinstance(cls, QuerySet) else cls
-
-    return inspect(
-        df=_filter_df_based_on_species(orm=cls, species=kwargs.get("species")),
-        identifiers=identifiers,
-        field=str(field),
-        case_sensitive=case_sensitive,
-        inspect_synonyms=inspect_synonyms,
-        return_df=return_df,
-        logging=logging,
-    )
-
-
-@classmethod  # type: ignore
-@doc_args(Registry.inspect.__doc__)
-def inspect(
-    cls,
-    identifiers: ListLike,
-    field: StrField,
-    *,
-    case_sensitive: bool = False,
-    inspect_synonyms: bool = True,
-    return_df: bool = False,
-    logging: bool = True,
-    **kwargs,
-) -> Union["pd.DataFrame", Dict[str, List[str]]]:
-    """{}"""
-    return _inspect(
-        cls=cls,
-        identifiers=identifiers,
-        field=field,
-        case_sensitive=case_sensitive,
-        inspect_synonyms=inspect_synonyms,
-        return_df=return_df,
-        logging=logging,
-        **kwargs,
-    )
-
-
-def _map_synonyms(
-    cls,
-    synonyms: Iterable,
-    *,
-    return_mapper: bool = False,
-    case_sensitive: bool = False,
-    keep: Literal["first", "last", False] = "first",
-    synonyms_field: str = "synonyms",
-    field: Optional[str] = None,
-    **kwargs,
-) -> Union[List[str], Dict[str, str]]:
-    """{}"""
-    from lamin_utils._map_synonyms import map_synonyms
-
-    if field is None:
-        field = get_default_str_field(cls)
-    if not isinstance(field, str):
-        field = field.field.name
-
-    cls = cls.model if isinstance(cls, QuerySet) else cls
-
-    try:
-        cls._meta.get_field(synonyms_field)
-        df = _filter_df_based_on_species(orm=cls, species=kwargs.get("species"))
-    except FieldDoesNotExist:
-        df = pd.DataFrame()
-    return map_synonyms(
-        df=df,
-        identifiers=synonyms,
-        field=field,
-        return_mapper=return_mapper,
-        case_sensitive=case_sensitive,
-        keep=keep,
-        synonyms_field=synonyms_field,
-    )
-
-
-@classmethod  # type: ignore
-@doc_args(Registry.map_synonyms.__doc__)
-def map_synonyms(
-    cls,
-    synonyms: Iterable,
-    *,
-    return_mapper: bool = False,
-    case_sensitive: bool = False,
-    keep: Literal["first", "last", False] = "first",
-    synonyms_field: str = "synonyms",
-    field: Optional[str] = None,
-    **kwargs,
-) -> Union[List[str], Dict[str, str]]:
-    """{}"""
-    return _map_synonyms(
-        cls=cls,
-        synonyms=synonyms,
-        return_mapper=return_mapper,
-        case_sensitive=case_sensitive,
-        keep=keep,
-        synonyms_field=synonyms_field,
-        field=field,
-        **kwargs,
-    )
-
-
 def describe(self):
     model_name = colors.green(self.__class__.__name__)
     msg = ""
@@ -529,33 +410,6 @@ def describe(self):
     settings.verbosity = verbosity
 
 
-def set_abbr(self, value: str):
-    try:
-        self.add_synonym(value, save=False)
-    except NotImplementedError:
-        pass
-    self.abbr = value
-    if not self._state.adding:
-        self.save()
-
-
-def _filter_df_based_on_species(
-    orm: Union[Registry, QuerySet], species: Optional[Union[str, Registry]] = None
-):
-    import pandas as pd
-
-    records = orm.all() if isinstance(orm, QuerySet) else orm.objects.all()
-    if _has_species_field(orm):
-        # here, we can safely import lnschema_bionty
-        from lnschema_bionty._bionty import create_or_get_species_record
-
-        species_record = create_or_get_species_record(species=species, orm=orm)
-        if species_record is not None:
-            records = records.filter(species__name=species_record.name)
-
-    return pd.DataFrame.from_records(records.values())
-
-
 def get_default_str_field(orm: Union[Registry, QuerySet, Manager]) -> str:
     """Get the 1st char or text field from the orm."""
     if isinstance(orm, (QuerySet, Manager)):
@@ -584,115 +438,12 @@ def get_default_str_field(orm: Union[Registry, QuerySet, Manager]) -> str:
     return field.name
 
 
-def _add_or_remove_synonyms(
-    synonym: Union[str, Iterable],
-    record: Registry,
-    action: Literal["add", "remove"],
-    force: bool = False,
-    save: Optional[bool] = None,
-):
-    """Add or remove synonyms."""
-
-    def check_synonyms_in_all_records(synonyms: Set[str], record: Registry):
-        """Errors if input synonym is associated with other records in the DB."""
-        import pandas as pd
-        from IPython.display import display
-
-        syns_all = (
-            record.__class__.objects.exclude(synonyms="").exclude(synonyms=None).all()
-        )
-        if len(syns_all) == 0:
-            return
-        df = pd.DataFrame(syns_all.values())
-        df["synonyms"] = df["synonyms"].str.split("|")
-        df = df.explode("synonyms")
-        matches_df = df[(df["synonyms"].isin(synonyms)) & (df["id"] != record.id)]
-        if matches_df.shape[0] > 0:
-            records_df = pd.DataFrame(syns_all.filter(id__in=matches_df["id"]).values())
-            logger.error(
-                f"Input synonyms {matches_df['synonyms'].unique()} already associated"
-                " with the following records:\n(Pass `force=True` to ignore this error)"
-            )
-            display(records_df)
-            raise SystemExit(AssertionError)
-
-    # passed synonyms
-    if isinstance(synonym, str):
-        syn_new_set = set([synonym])
-    else:
-        syn_new_set = set(synonym)
-    # nothing happens when passing an empty string or list
-    if len(syn_new_set) == 0:
-        return
-    # because we use | as the separator
-    if any(["|" in i for i in syn_new_set]):
-        raise AssertionError("A synonym can't contain '|'!")
-
-    # existing synonyms
-    syns_exist = record.synonyms
-    if syns_exist is None or len(syns_exist) == 0:
-        syns_exist_set = set()
-    else:
-        syns_exist_set = set(syns_exist.split("|"))
-
-    if action == "add":
-        if not force:
-            check_synonyms_in_all_records(syn_new_set, record)
-        syns_exist_set.update(syn_new_set)
-    elif action == "remove":
-        syns_exist_set = syns_exist_set.difference(syn_new_set)
-
-    if len(syns_exist_set) == 0:
-        syns_str = None
-    else:
-        syns_str = "|".join(syns_exist_set)
-
-    record.synonyms = syns_str
-
-    if save is None:
-        # if record is already in DB, save the changes to DB
-        save = not record._state.adding
-    if save:
-        record.save()
-
-
-def _check_synonyms_field_exist(record: Registry):
-    try:
-        record.__getattribute__("synonyms")
-    except AttributeError:
-        raise NotImplementedError(
-            f"No synonyms field found in table {record.__class__.__name__}!"
-        )
-
-
-def add_synonym(
-    self,
-    synonym: Union[str, ListLike],
-    force: bool = False,
-    save: Optional[bool] = None,
-):
-    _check_synonyms_field_exist(self)
-    _add_or_remove_synonyms(
-        synonym=synonym, record=self, force=force, action="add", save=save
-    )
-
-
-def remove_synonym(self, synonym: Union[str, ListLike]):
-    _check_synonyms_field_exist(self)
-    _add_or_remove_synonyms(synonym=synonym, record=self, action="remove")
-
-
 METHOD_NAMES = [
     "__init__",
     "search",
     "lookup",
-    "map_synonyms",
-    "inspect",
-    "add_synonym",
-    "remove_synonym",
     "from_values",
     "describe",
-    "set_abbr",
     "view_parents",
 ]
