+++ pyproject.toml
@@ -19,7 +19,7 @@ dependencies = [
     # !!! They cannot be pinned ANYWHERE else !!!
     "lnschema_core==0.43.3",
     "lamindb_setup==0.49.4",
-    "lamin_utils==0.9.3",
+    "lamin_utils==0.9.4",
     # Lamin GREATEREQ packages
     "erdiagram>=0.1.2",
     # others
